+++ extensions/easydict/src/releaseVersion/changelog.ts
@@ -2,25 +2,30 @@
  * @author: tisfeng
  * @createTime: 2022-07-03 22:10
  * @lastEditor: tisfeng
- * @lastEditTime: 2022-07-16 23:24
+ * @lastEditTime: 2022-07-20 21:17
  * @fileName: changelog.ts
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.
  */
 
 export const changelog = `
-## [1.3.0] - 2022-07-16
+## [1.3.0] - 2022-07-20
 
 ### ✨ 新特性
 
 - 支持 DeepL 翻译。
 - 支持手动排序翻译结果显示。
 
-![easydict-5](https://raw.githubusercontent.com/tisfeng/ImageBed/c393f8bbb9e4a332e0058854494a08bd1e1bdc08/uPic/easydict-5.png)
+![easydict-5](https://raw.githubusercontent.com/tisfeng/ImageBed/main/uPic/easydict-5-1658309785.png)
+
+### 💎 Improvements
+
+- 使用了一些新的 icon 去适配 raycast 新 API。
 
 ### 🐞 Fixes
 
-- 修复删除输入框文本会导致重复查询选中文本问题。
+- 修复了删除输入框文本会导致重复查询选中文本问题。
+- 修复了动作按钮渲染时会闪烁显示 “✨ New Version Released” 问题。
 
 ### ✅ 最近更新
 
@@ -35,9 +40,14 @@ export const changelog = `
 - Added support for DeepL translation.
 - Supported to sort translation results manually.
 
+### 💎 Improvements
+
+- Used the new icons to adapted to the new raycast API.
+
 ### 🐞 Fixes
 
 - Fixed deleting input box text will cause repeated query selected text problem.
+- Fixed the bug of "✨ New Version Released" flashing when rendering action button.
 
 ### ✅ Recent Updates