+++ extensions/easydict/src/releaseVersion/versionInfo.ts
@@ -2,7 +2,7 @@
  * @author: tisfeng
  * @createTime: 2022-07-01 19:05
  * @lastEditor: tisfeng
- * @lastEditTime: 2022-07-17 10:37
+ * @lastEditTime: 2022-07-20 02:00
  * @fileName: versionInfo.ts
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.
@@ -10,7 +10,6 @@
 
 import { LocalStorage } from "@raycast/api";
 import axios from "axios";
-import { requestCostTime } from "../request";
 import { changelog } from "./changelog";
 
 const versionInfoKey = "EasydictVersionInfoKey";
@@ -143,7 +142,7 @@ export class Easydict {
     try {
       // console.log(`fetch release url: ${releaseUrl}`);
       const response = await axios.get(releaseUrl);
-      console.log(`fetch github cost time: ${response.headers[requestCostTime]} ms`);
+      console.log(`fetch github cost time: ${response.headers["x-request-cost"]} ms`);
 
       return Promise.resolve(response.data);
     } catch (error) {